Add faster hashing for booleans/numbers/pointers
- Add table.cpp:floor_log2() to help in determining which indices to use
  in `index_ranges`
- Add name of shadowed local when throwing in parser.cpp:local_push()
- Add '_' as an exception to local shadowing
- Remove load factor for string table
- Add string linked-list printing for `Object *` pretty-printer
- Change type of all `pc` variables from isize to int

+++ cpp/src/debug.hpp
@@ -9,7 +9,7 @@ int
 debug_get_pad(const Chunk *p);
 
 void
-debug_disassemble_at(const Chunk *p, Instruction ip, isize index, int pad);
+debug_disassemble_at(const Chunk *p, Instruction ip, int index, int pad);
 
 [[noreturn]]
 void
